Simplify IncludeFile action
