+++ front/lib/actions/runners.ts
@@ -1,5 +1,6 @@
 import type { BrowseConfigurationType } from "@app/lib/actions/browse";
 import { BrowseConfigurationServerRunner } from "@app/lib/actions/browse";
+import type { ConversationIncludeFileConfigurationType } from "@app/lib/actions/conversation/include_file";
 import { ConversationIncludeFileConfigurationServerRunner } from "@app/lib/actions/conversation/include_file";
 import type { DustAppRunConfigurationType } from "@app/lib/actions/dust_app_run";
 import { DustAppRunConfigurationServerRunner } from "@app/lib/actions/dust_app_run";
@@ -19,7 +20,6 @@ import type {
   BaseActionConfigurationServerRunnerConstructor,
   BaseActionConfigurationStaticMethods,
 } from "@app/lib/actions/types/base";
-import type { ConversationIncludeFileConfigurationType } from "@app/lib/actions/types/conversation/include_file";
 import type { WebsearchConfigurationType } from "@app/lib/actions/websearch";
 import { WebsearchConfigurationServerRunner } from "@app/lib/actions/websearch";
